DictionarySourceCoordinator update interface
