+++ src/Dictionaries/DictionarySource.h
@@ -1,12 +1,8 @@
 #pragma once
 
 #include <memory>
-#include <Columns/ColumnDecimal.h>
-#include <Columns/ColumnString.h>
-#include <Columns/ColumnVector.h>
 #include <Columns/IColumn.h>
 #include <Core/Names.h>
-#include <DataTypes/DataTypesNumber.h>
 #include <Processors/Sources/SourceWithProgress.h>
 #include <Dictionaries/DictionaryStructure.h>
 #include <Dictionaries/IDictionary.h>
@@ -15,9 +11,17 @@
 namespace DB
 {
 
-class DictionarySourceCoordinator
+class DictionarySource;
+
+class DictionarySourceCoordinator final : public shared_ptr_helper<DictionarySourceCoordinator>, public std::enable_shared_from_this<DictionarySourceCoordinator>
 {
+    friend struct shared_ptr_helper<DictionarySourceCoordinator>;
+
 public:
+
+    Pipe read(size_t num_streams);
+
+private:
 
     explicit DictionarySourceCoordinator(
         std::shared_ptr<const IDictionary> dictionary_,
@@ -45,6 +49,8 @@ public:
         initialize(column_names);
     }
 
+    friend class DictionarySource;
+
     bool getKeyColumnsNextRangeToRead(ColumnsWithTypeAndName & key_columns, ColumnsWithTypeAndName & data_columns);
 
     const Block & getHeader() const { return header; }
@@ -57,7 +63,6 @@ public:
 
     const std::shared_ptr<const IDictionary> & getDictionary() const { return dictionary; }
 
-private:
     void initialize(const Names & column_names);
 
     static ColumnsWithTypeAndName cutColumns(const ColumnsWithTypeAndName & columns_with_type, size_t start, size_t length);
@@ -75,23 +80,6 @@ private:
 
     const size_t max_block_size;
     std::atomic<size_t> parallel_read_block_index = 0;
-};
-
-class DictionarySource : public SourceWithProgress
-{
-public:
-
-    explicit DictionarySource(std::shared_ptr<DictionarySourceCoordinator> coordinator_)
-        : SourceWithProgress(coordinator_->getHeader()), coordinator(std::move(coordinator_))
-    {
-    }
-
-private:
-    String getName() const override { return "DictionarySource"; }
-
-    Chunk generate() override;
-
-    std::shared_ptr<DictionarySourceCoordinator> coordinator;
 };
 
 }
